backend/requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pymysql
python-dotenv
pydantic[email]
cors
